moshop-修改配置，下单支付全链路多场景测试准备，修复bug-购物车总价格无法显示

+++ mo-order-service/src/main/java/com/mo/component/AlipayStrategy.java
@@ -60,7 +60,7 @@ public class AlipayStrategy implements PayStrategy {
         }
 
         // 该笔订单允许的最晚付款时间，逾期将关闭交易。取值范围：1m～15d。m-分钟，h-小时，d-天，1c-当天（1c-当天的情况下，无论交易何时创建，都在0点关闭）。 该参数数值不接受小数点， 如 1.5h，可转换为 90m。
-        content.put("timeout_express", Double.valueOf(timeout) + "m");
+        content.put("timeout_express", Double.valueOf(timeout).intValue() + "m");
 
         //端类型
         String clientType = payInfoVO.getClientType();
